+++ source/frontend/buttons/bbcomplete.js
@@ -116,7 +116,7 @@ module.exports = new ButtonWrapper(mainId, 3000,
 				if (company.scoreboardIsSeasonal) {
 					embeds.push(await seasonalScoreboardEmbed(company, collectedInteraction.guild, participationMap, descendingRanks, goalProgress));
 				} else {
-					embeds.push(await overallScoreboardEmbed(company, collectedInteraction.guild, await logicLayer.hunters.findCompanyHunters(collectedInteraction.guild.id), descendingRanks, goalProgress));
+					embeds.push(await overallScoreboardEmbed(company, collectedInteraction.guild, await logicLayer.hunters.findCompanyHunters(collectedInteraction.guild.id), goalProgress));
 				}
 				updateScoreboard(company, collectedInteraction.guild, embeds);
 			}).catch(error => {