+++ source/frontend/commands/bounty/complete.js
@@ -114,7 +114,7 @@ module.exports = new SubcommandWrapper("complete", "Close one of your open bount
 			if (company.scoreboardIsSeasonal) {
 				embeds.push(await seasonalScoreboardEmbed(company, interaction.guild, participationMap, descendingRanks, goalProgress));
 			} else {
-				embeds.push(await overallScoreboardEmbed(company, interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), descendingRanks, goalProgress));
+				embeds.push(await overallScoreboardEmbed(company, interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), goalProgress));
 			}
 			updateScoreboard(company, interaction.guild, embeds);
 		});