+++ source/frontend/commands/bounty/post.js
@@ -186,7 +186,7 @@ module.exports = new SubcommandWrapper("post", "Post your own bounty (+1 XP)",
 					if (company.scoreboardIsSeasonal) {
 						embeds.push(await seasonalScoreboardEmbed(company, interaction.guild, participationMap, descendingRanks, goalProgress));
 					} else {
-						embeds.push(await overallScoreboardEmbed(company, interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), descendingRanks, goalProgress));
+						embeds.push(await overallScoreboardEmbed(company, interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), goalProgress));
 					}
 					updateScoreboard(company, interaction.guild, embeds);
 				});