+++ source/frontend/commands/reset/all-hunter-stats.js
@@ -12,12 +12,11 @@ module.exports = new SubcommandWrapper("all-hunter-stats", "IRREVERSIBLY reset a
 			await logicLayer.seasons.deleteSeasonParticipations(season.id);
 		}
 		const embeds = [];
-		const ranks = await logicLayer.ranks.findAllRanks(interaction.guild.id);
 		const goalProgress = await logicLayer.goals.findLatestGoalProgress(interaction.guild.id);
 		if (company.scoreboardIsSeasonal) {
-			embeds.push(await seasonalScoreboardEmbed(company, interaction.guild, new Map(), ranks, goalProgress));
+			embeds.push(await seasonalScoreboardEmbed(company, interaction.guild, new Map(), await logicLayer.ranks.findAllRanks(interaction.guild.id), goalProgress));
 		} else {
-			embeds.push(await overallScoreboardEmbed(company, interaction.guild, [], ranks, goalProgress));
+			embeds.push(await overallScoreboardEmbed(company, interaction.guild, [], goalProgress));
 		}
 		updateScoreboard(company, interaction.guild, embeds);
 		interaction.user.send(`Resetting bounty hunter stats on ${interaction.guild.name} has completed.`);