+++ source/frontend/commands/toast.js
@@ -115,7 +115,7 @@ module.exports = new CommandWrapper(mainId, "Raise a toast to other bounty hunte
 				if (company.scoreboardIsSeasonal) {
 					embeds.push(await seasonalScoreboardEmbed(company, interaction.guild, participationMap, descendingRanks, goalProgress));
 				} else {
-					embeds.push(await overallScoreboardEmbed(company, interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), descendingRanks, goalProgress));
+					embeds.push(await overallScoreboardEmbed(company, interaction.guild, await logicLayer.hunters.findCompanyHunters(interaction.guild.id), goalProgress));
 				}
 				updateScoreboard(company, interaction.guild, embeds);
 			}